+++ lib/js_routes/generators/middleware.rb
@@ -1,25 +1,25 @@
-require "rails/generators"
+require "js_routes/generators/base"
 
-class JsRoutes::Generators::Middleware < Rails::Generators::Base
-
-  source_root File.expand_path(__FILE__ + "/../../../templates")
+class JsRoutes::Generators::Middleware < JsRoutes::Generators::Base
 
   def create_middleware
     copy_file "initializer.rb", "config/initializers/js_routes.rb"
-    inject_into_file "app/javascript/packs/application.js", pack_content
     inject_into_file "config/environments/development.rb", middleware_content, before: /^end\n\z/
     inject_into_file "Rakefile", rakefile_content
     inject_into_file ".gitignore", gitignore_content
+    if path = application_js_path
+      inject_into_file path, pack_content
+    end
     JsRoutes.generate!
-    JsRoutes.definitions!
+    JsRoutes.definitions! if JsRoutes.configuration.modern?
   end
 
   protected
 
   def pack_content
     <<-JS
-import * as Routes from '../routes';
-window.Routes = Routes;
+import {root_path} from '../routes';
+alert(`JsRoutes installed.\\nYour root path is ${root_path()}`)
     JS
   end
 
@@ -33,12 +33,11 @@ window.Routes = Routes;
   end
 
   def rakefile_content
+    enhanced_task = depends_on_js_bundling? ? "javascript:build" : "assets:precompile"
+    dependency_task = JsRoutes.configuration.modern? ? "js:routes:typescript" : 'js:routes'
     <<-RB
-
-# Update js-routes file before assets precompile
-namespace :assets do
-  task :precompile => "js:routes:typescript"
-end
+# Update js-routes file before javascript build
+task "#{enhanced_task}" => "#{dependency_task}"
     RB
   end