+++ lib/js_routes/generators/webpacker.rb
@@ -2,14 +2,14 @@ require "rails/generators"
 
 class JsRoutes::Generators::Webpacker < Rails::Generators::Base
 
-  source_root File.expand_path(__FILE__ + "/../../../templates")
-
   def create_webpack
     copy_file "initializer.rb", "config/initializers/js_routes.rb"
     copy_file "erb.js", "config/webpack/loaders/erb.js"
-    copy_file "routes.js.erb", "app/javascript/routes.js.erb"
+    copy_file "routes.js.erb", "#{Webpacker.config.source_path}/routes.js.erb"
     inject_into_file "config/webpack/environment.js", loader_content
-    inject_into_file "app/javascript/packs/application.js", pack_content
+    if path = application_js_path
+      inject_into_file path, pack_content
+    end
     command = Rails.root.join("./bin/yarn add rails-erb-loader")
     run command
   end
@@ -19,7 +19,7 @@ class JsRoutes::Generators::Webpacker < Rails::Generators::Base
   def pack_content
     <<-JS
 import * as Routes from 'routes.js.erb';
-window.Routes = Routes;
+alert(Routes.root_path())
     JS
   end