+++ spec/js_routes/module_types/dts_spec.rb
@@ -6,13 +6,19 @@ require "spec_helper"
 
 describe JsRoutes, "compatibility with DTS"  do
 
-  OPTIONS = {module_type: 'DTS', include: [/^inboxes$/, /^inbox_message_attachment$/]}
+  INCLUDE_OPTION = [/^inboxes$/, /^inbox_message_attachment$/]
+  OPTIONS = {module_type: 'DTS', include: INCLUDE_OPTION}
+
   let(:extra_options) do
     {}
   end
 
+  let(:options) do
+    { module_type: 'DTS', include: INCLUDE_OPTION, **extra_options }
+  end
+
   let(:generated_js) do
-    JsRoutes.generate({**OPTIONS, **extra_options})
+    JsRoutes.generate(**options)
   end
 
   context "when file is generated" do
@@ -46,7 +52,7 @@ describe JsRoutes, "compatibility with DTS"  do
       expect(generated_js).to include(<<-DOC.rstrip)
 /**
  * Generates rails route to
- * /inboxes/:inbox_id/messages/:message_id/attachments/:id(.:format)
+ * /inboxes/:inbox_id/messages/:message_id/attachments/:id
  * @param {any} inboxId
  * @param {any} messageId
  * @param {any} id
@@ -57,7 +63,7 @@ export const inboxMessageAttachmentPath: ((
   inboxId: RequiredRouteParameter,
   messageId: RequiredRouteParameter,
   id: RequiredRouteParameter,
-  options?: {format?: OptionalRouteParameter} & RouteOptions
+  options?: RouteOptions
 ) => string) & RouteHelperExtras;
 DOC
     end
@@ -78,7 +84,7 @@ DOC
     expect(generated_js).to include(<<-DOC.rstrip)
 /**
  * Generates rails route to
- * /inboxes/:inbox_id/messages/:message_id/attachments/:id(.:format)
+ * /inboxes/:inbox_id/messages/:message_id/attachments/:id
  * @param {any} inbox_id
  * @param {any} message_id
  * @param {any} id
@@ -89,7 +95,7 @@ export const inbox_message_attachment_path: ((
   inbox_id: RequiredRouteParameter,
   message_id: RequiredRouteParameter,
   id: RequiredRouteParameter,
-  options?: {format?: OptionalRouteParameter} & RouteOptions
+  options?: RouteOptions
 ) => string) & RouteHelperExtras
 DOC
   end
@@ -105,7 +111,16 @@ DOC
   describe "compiled javascript asset" do
     subject { ERB.new(File.read("app/assets/javascripts/js-routes.js.erb")).result(binding) }
     it "should have js routes code" do
-      is_expected.to include("export const inbox_message_path = __jsr.r(")
+      is_expected.to include("export const inbox_message_path = /*#__PURE__*/ __jsr.r(")
+    end
+  end
+
+  describe ".definitions" do
+    let(:extra_options) { { module_type: 'ESM' } }
+
+    it "uses DTS module automatically" do
+      generated_js = JsRoutes.definitions(**options)
+      expect(generated_js).to include('export {};')
     end
   end
 end