+++ spec/js_routes/zzz_after_initialization_spec.rb
@@ -2,14 +2,21 @@
 # because it cause unrevertable changes to runtime
 # what is why I added "zzz_last" in the beginning
 
+require "sprockets"
+require "sprockets/railtie"
 require 'spec_helper'
 require "fileutils"
 
 describe "after Rails initialization", :slow do
   NAME = Rails.root.join('app', 'assets', 'javascripts', 'routes.js').to_s
+  CONFIG_ROUTES = Rails.root.join('config','routes.rb').to_s
 
   def sprockets_v3?
     Sprockets::VERSION.to_i >= 3
+  end
+
+  def sprockets_v4?
+    Sprockets::VERSION.to_i >= 4
   end
 
   def sprockets_context(environment, name, filename)
@@ -30,17 +37,18 @@ describe "after Rails initialization", :slow do
 
   before(:each) do
     FileUtils.rm_rf Rails.root.join('tmp/cache')
-    FileUtils.rm_f NAME
+    JsRoutes.remove!(NAME)
     JsRoutes.generate!(NAME)
   end
 
   before(:all) do
+    JsRoutes::Engine.install_sprockets!
     Rails.configuration.eager_load = false
     Rails.application.initialize!
   end
 
   it "should generate routes file" do
-    expect(File.exists?(NAME)).to be_truthy
+    expect(File.exist?(NAME)).to be_truthy
   end
 
   it "should not rewrite routes file if nothing changed" do
@@ -58,6 +66,28 @@ describe "after Rails initialization", :slow do
     expect(File.mtime(NAME)).not_to eq(routes_file_mtime)
   end
 
+  describe JsRoutes::Middleware do
+
+    it "works" do
+      JsRoutes.remove!
+      expect(File.exists?(NAME)).to be(false)
+      app = lambda do |env|
+        [200, {}, ""]
+      end
+      middleware = JsRoutes::Middleware.new(app)
+      middleware.call({})
+
+      expect(File.exists?(NAME)).to be(true)
+      JsRoutes.remove!
+      middleware.call({})
+      expect(File.exists?(NAME)).to be(false)
+      FileUtils.touch(CONFIG_ROUTES)
+      middleware.call({})
+      expect(File.exists?(NAME)).to be(true)
+    end
+
+  end
+
   context "JsRoutes::Engine" do
     TEST_ASSET_PATH = Rails.root.join('app','assets','javascripts','test.js')
 
@@ -72,11 +102,10 @@ describe "after Rails initialization", :slow do
 
     context "the preprocessor" do
       before(:each) do
-        path = Rails.root.join('config','routes.rb').to_s
-        if sprockets_v3?
-          expect_any_instance_of(Sprockets::Context).to receive(:depend_on).with(path)
+        if sprockets_v3? || sprockets_v4?
+          expect_any_instance_of(Sprockets::Context).to receive(:depend_on).with(CONFIG_ROUTES.to_s)
         else
-          expect(ctx).to receive(:depend_on).with(path)
+          expect(ctx).to receive(:depend_on).with(CONFIG_ROUTES.to_s)
         end
       end
       let!(:ctx) do
@@ -86,7 +115,6 @@ describe "after Rails initialization", :slow do
       end
 
       context "when dealing with js-routes.js" do
-
         context "with Rails" do
           context "and initialize on precompile" do
             before(:each) do